Rename story with final name
