+++ packages/stickers/src/baking-bread-guide/flourBag.js
