+++ packages/stickers/src/baking-bread-guide/flourBowl.js
