+++ packages/stickers/src/baking-bread-guide/thermoStat.js
